common/kupumultieditor.js: Removed "focus" handler that called _clearSelection().
common/kupuhelpers.js: Fixed BaseSelection.selectionInsideNode() so that it 
works as advertised.

common/kupueditor.js: Femoved some spurious calls to _restoreSelection();
getSelection() does this for you. Removed "focus" handler that calls
_clearSelection(); _restoreSelection() will correctly not restore a saved
selection if the current document is the focus. Made _saveSelection() check
if the current selection is in the document before saving it.