kupuploneui.js: Another fix for setting styles in IE: when removing class need removeAttribute("className") on IE, removeAttribute("class") on Mozilla.
kupu files: Added $Id$ keyword to all the .kupu files. Modified the xslt to copy the ids into comments at the top of the Plone macro. This should make it possible to write an automated check that the make has been run and is up to date whenever Zope starts.