Implement kssaction wrapper for action view methods
This is an alternate way of making view methods. First, self.render()
is not needed to be returned at the end. Second, if KssExplicitError
is raised, that will not cause an error log on the server, but
instead it sends the error command to the client, which in return
displays the fallback reason and executes its error action.