Started to make as much tests for the kss.core compatibility layer work
The meta.zcml now includes the meta.zcml from kss.zope, this makes
sure this is always loaded

Readded the siteview test. This is a duplicate of the one in kss.zope
but since the old kss.core had it the first version of kss.core should
have it as well (to maximize coverage).