Added some of the changes from the diff file for issue #9. Made some changes to allow the attached image on the video display window to take the same size as the video so the page would at least be useable with a large image (issue#5). Although technically a "feature" the data added from the diff was easily used to make issue #5 less urgent.
The python code gives a default size to the video of "320x240" (if it
is not passed  a size). This is done anyways in the js file, but was
added here to give a default to the image as well. 

The formats tested were mov, swf, mpg, wmv. I did not get a chance to
test the real-media as of yet.