Begin re-thinking the differential builder around labels.
The idea will be to use labels to determine the axes for multiple
differential reports.  This allows for better user direction in terms of
generating the results and/or grooming the reports directory.  It also
then allows multiple tests to be meaningfully integrated into one table
of differential reports.

This also changes the way that continuous integration/buildbot will
work.