Adding handling of None values
In zope.formlib, when one asks for entered data, you get a dictionary
with values for all fields of an interface. If the values are not
required, the value is None. Propertysheets might stumble over None
values therefor we don't set them on propertysheets.
Otherwise this breaks LDAP User Management