Refactored the tests into unit tests and integration tests.
It is not sensible to modify the CA registrations in integration tests because
this breaks the assumptions made by plone(.app).testing layer setup. The idea
in plone(.app).testing is to set up the CA in particular manner for a layer
which is a shared state between all tests within the layer. Modifying the CA
registrations within a single test breaks this contract.

We can also see that the raptus article registrations are not in the global
registry nor the local registry by inspecting

    zope.component.globalSiteManager._adapter_registrations

nor are they in the Plone local site manager

    self.portal.getSiteManager()._adapter_registrations

which is why the self.portal.getSiteManager().unregisterAdapter() call failed.
They are registered in a custom site manager set up by plone.app.testing which
can be seen by inspecting

    self.portal.getSiteManager().__bases__[0]._adapter_registrations

Calling

    self.portal.getSiteManager().__bases__[0].unregisterAdapter(...)

would unregister the adapter as was the original attempt but this
unregistering would leak to all tests in the same layer that run after the
call to unregisterAdapter() therefore breaking the layer contract.

In these cases I find the better to test the edge cases of the functions in
unit tests and mock the CA since we want to test the functionality itself, not
the Zope CA. We can then use integration tests for positive assertations.

For mocking, the mock package has proven very useful in testing. See

    http://www.voidspace.org.uk/python/mock/

for details.