respect Site Setup/Security/"Use email address as login name" by merging
    http://svn.plone.org/svn/collective/Products.remember/branches/maurits-emaillogin-pas

which integrates a PAS auth plugin for email login ids.  i add a bit more
which activates the PAS plugin when Products.remember is installed (rather
than requiring manual activation in the PAS tool), and also provides an
upgrade step for activating the plugin in portals which already use
Products.rmemember (in configure.zcml, which i'll check in next).

New accounts are not allowed to have the same email address as already
existing ones when "Use email address as login name" is activated, but the
policy is not enforced when it is deactivated.  (Already existing accounts
are not subject to the constraint in either case.)